refactor(GenericFactory): moves factory to `generic` package
